// src/app/pages/solicitudewarehouse_products/components/solicitudewarehouse_products-table/solicitudewarehouse_products.interface.ts
export interface Solicitudewarehouse_productsInterface {
   idsolicitudewarehouse_product?: number;
   solicitudewarehouse_idsolicitudewarehouse?: number;
   product_idproduct?: number;
   quantity?: number;
   is_deleted?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}